+++ Sources/Usd/Classes/TimeCode.swift
@@ -11,7 +11,7 @@ import SwiftPy
 /// Represent a time value, which may be either numeric, holding a double value, or a sentinel value TimeCode.Default()
 @Scriptable(convertsToSnakeCase: false)
 public class TimeCode: ClassWrapper<pxr.UsdTimeCode> {
-    public init(_ time: Double? = nil) {
+    public init(_ time: Int? = nil) {
         if let time {
             super.init(value: UsdTimeCode(time))
         } else {
